Langgraph/Capstone_Langgraph_Project/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
tavily-python
python-dotenv
pandas
requests
redis